# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Beatboxd (Letterboxd for DJ sets). Generated from\nserver/internal/httpapi handler annotations via swag - run\n`make openapi` in server/ to regenerate after changing a handler.",
        "title": "Beatboxd API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/follows/{userId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Follow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id to follow",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Unfollow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id to unfollow",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Leaderboard scoped to the caller and everyone they follow",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LeaderboardEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get the caller's own user row",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update the caller's own profile",
                "parameters": [
                    {
                        "description": "fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/users/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Search users by username or display name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{userId}/followers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "List the users who follow the given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{userId}/following": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "List the users the given user follows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{username}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user's profile by username, with follow/review counts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserProfileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/users/{username}/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user's review stats: totals, top DJs, top venues",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserStatsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Dj": {
            "type": "object",
            "required": [
                "bio",
                "createdAt",
                "createdByUserId",
                "genres",
                "id",
                "imageUrl",
                "name",
                "slug",
                "spotifyId",
                "updatedAt"
            ],
            "properties": {
                "bio": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "spotifyId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "LeaderboardEntry": {
            "type": "object",
            "required": [
                "reviewCount",
                "user"
            ],
            "properties": {
                "reviewCount": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
...
        "SuccessResponse": {
            "type": "object",
            "required": [
                "success"
            ],
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "TopDj": {
            "type": "object",
            "required": [
                "dj",
                "reviewCount"
            ],
            "properties": {
                "dj": {
                    "$ref": "#/definitions/Dj"
                },
                "reviewCount": {
                    "type": "integer"
                }
            }
        },
        "TopVenue": {
            "type": "object",
            "required": [
                "reviewCount",
                "venue"
            ],
            "properties": {
                "reviewCount": {
                    "type": "integer"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "required": [
                "avatarUrl",
                "bio",
                "createdAt",
                "displayName",
                "id",
                "updatedAt",
                "username"
            ],
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "UserProfileResponse": {
            "type": "object",
            "required": [
                "followerCount",
                "followingCount",
                "reviewCount",
                "user"
            ],
            "properties": {
                "followerCount": {
                    "type": "integer"
                },
                "followingCount": {
                    "type": "integer"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
        "UserStatsResponse": {
            "type": "object",
            "required": [
                "topDjs",
                "topVenues",
                "totalReviews",
                "uniqueDjs"
            ],
            "properties": {
                "topDjs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TopDj"
                    }
                },
                "topVenues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TopVenue"
                    }
                },
                "totalReviews": {
                    "type": "integer"
                },
                "uniqueDjs": {
                    "type": "integer"
                }
            }
        },
...
        "apiError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "errorEnvelope": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/apiError"
                }
            }
        },
...
        "updateProfileRequest": {
            "type": "object",
            "required": [
                "avatarUrl",
                "bio",
                "displayName",
                "username"
            ],
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Clerk-issued session JWT, e.g. \"Bearer eyJhbGc...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}